+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added `balm.m.object` module for additional object helper functions
   * Added `balm.m.object.construct/2` for reconstructing an object from a simple table or instance
   * Added `balm.m.object.construct_record/2` for constructing a record of objects from a table
+* Added `balm.m.vector2.to_hash/2`
+* Added `balm.m.vector2.from_hash/3`
 
 # 2026.5.4